minor fixes- grammar, paths, file names etc
Comment: with resources.path function doesn't seem to work with "data" as input for me and I think may be deprecated so not sure if it is better to use the regular pathlib. I've changed this to pathlib in one location so up to you if to keep it or not!